Connected Feature Frame and Feature service to the measurements and descriptions.  Also sorted out bugs in the Measurement Processor that was generating a poorly formatted arraylist.
NOTE:  All processes need to be validated still.

+++ src/main/java/vtea/objects/layercake/microVolume.java
@@ -689,7 +689,7 @@ public void addRegions(List<microRegion> regions){
     }
 
     @Override
-    public int getSerialID() {
+    public double getSerialID() {
        return serialID;
     }
